# openapi.yaml
openapi: 3.0.3
info:
  title: Alauda API
  version: 1.0.0
  description: |
    🚀 **Alauda API - Multi-Service Integration Platform**
    
    All-in-one API for media downloads, social media content, music identification, payment processing and more. Built for developers who need reliable, fast and affordable integrations.
    
    ## 🎯 Features
    
    **Media Downloads:**
    - 🎵 Spotify - Search and download tracks
    - 📹 YouTube - Video and audio downloader
    - 🎬 TikTok - Download videos without watermark
    - 📸 Instagram - Download posts, reels, stories
    - 📘 Facebook - Video downloader
    
    **Music Services:**
    - 🎤 Shazam - Identify any song
    - 📝 Lyrics - Search song lyrics worldwide
    
    **Utilities:**
    - 🆔 CPF Validator - Brazilian tax ID validation
    - 🖼️ Background Remover - AI-powered image processing
    - 💬 WhatsApp Utils - Formatting and validation tools
    
    **Payment Processing:**
    - 💳 MercadoPago - Brazil and LATAM payments
    - 📱 M-Pesa - Mozambique mobile money (Vodacom)
    - 📱 E-Mola - Mozambique mobile money (Movitel)
    
    ## 🌍 Developed in Mozambique 🇲🇿 | Available Worldwide
    
    **Created by:** Zeus Lykraios 💎
    
    ## 🔑 Authentication
    
    All endpoints (except `/` and `/health`) require an API key sent in the `X-API-Key` header.
    
    Get your free API key at: https://alauda-api.topazioverse.com.br/api/auth/register
    
    ## 💰 Credits System
    
...
    
    ## 📊 Rate Limiting
    
...
    
  contact:
...
servers:
...
security:
  - ApiKeyAuth: []
paths:
...
  /api/admin/keys/add-credits:
    post:
      tags:
        - System
      summary: Add Credits to an API Key (Admin only)
      description: |
        Add credits (reqs) to an API key. RESTRICTED to the owner admin key
        (email: eliobrostech3@gmail.com). Send the admin key in the X-API-Key header.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - credits
              properties:
                apiKey:
                  type: string
                  description: Full target API key (alauda_live_...)
                  example: "alauda_live_abcdef..."
                email:
                  type: string
                  description: Owner email (uses the most recent active key)
                  example: "cliente@email.com"
                credits:
                  type: integer
                  minimum: 1
                  example: 500
                reason:
                  type: string
                  description: Optional reason recorded in recharge history
                  example: "bônus"
      responses:
        '200':
          description: Credits added
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      message:
                        type: string
                      credits_added:
                        type: integer
                      credits_total:
                        type: integer
                      target:
                        type: object
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Not the admin key
  /api/admin/keys:
    get:
      tags:
        - System
      summary: List API Keys (Admin only)
      description: List/search API keys. RESTRICTED to the owner admin key.
      parameters:
        - in: query
          name: email
          schema:
            type: string
        - in: query
          name: plan
          schema:
            type: string
        - in: query
          name: status
          schema:
            type: string
            enum: [active, suspended]
        - in: query
          name: limit
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: List of keys
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      count:
                        type: integer
                      total:
                        type: integer
                      keys:
                        type: array
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Not the admin key
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        API key for authentication. Get yours at: https://alauda-api.topazioverse.com.br/api/auth/register
  schemas:
...
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
        message:
          type: string
    ValidationError:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: "Validation error"
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
  responses:
    Unauthorized:
      description: Invalid or missing API key
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error: "API Key inválida ou expirada"
    ValidationError:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
...